Code refactoring, commented redundent code for now
